implement rounding setting

+++ ttime/Report.cs
@@ -8,11 +8,12 @@ namespace ttime
         public DateTime Start { get; set; }
         public DateTime End { get; set; }
         public List<Item> Items { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
 
-        public class Item {
+        public class Item
+        {
             public string Name { get; set; }
-            public double Hours { get; set; }
+            public decimal Hours { get; set; }
         }
     }
 }
